modularized the JavaScript pong game a bit
